requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM
openai>=1.3.0

# Async
aiohttp>=3.9.0
aiolimiter>=1.1.0
nest-asyncio>=1.5.0

# Database
aiosqlite>=0.19.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Analysis
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
pydantic>=2.0.0

